pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "backend template"
authors = ["Ebin Joseph <ebin7joseph@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
djangorestframework-jwt = "^1.11.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
